# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Cakra Prima",
        "description": "Dokumentasi API Kas & Keuangan Cakra Prima",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/potentials": {
            "get": {
                "tags": [
                    "Potentials"
                ],
                "summary": "Daftar Potensi",
                "operationId": "8de71b1322aa87e802bc38e3dfc4188b",
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Potentials"
                ],
                "summary": "Tambah Potensi",
                "operationId": "687d888aee82d9932fb6d402cbb327ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "description",
                                    "amount",
                                    "trans_date",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Proyek A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Potensi pendapatan bulan depan"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 5000000
                                    },
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-27"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil dibuat"
                    }
                }
            }
        },
        "/api/potentials/{id}": {
            "get": {
                "tags": [
                    "Potentials"
                ],
                "summary": "Detail Potensi",
                "operationId": "ac8476f64cf5d1fbd77acc770443679a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Potentials"
                ],
                "summary": "Update Potensi",
                "operationId": "c33f86da4ed62f9d3754b66582c629ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number"
                                    },
                                    "trans_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Potentials"
                ],
                "summary": "Hapus Potensi",
                "operationId": "bcc7292352bfb9652c238dbe2dc9033a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}